changed structure for resource mapper. added separate classes for etag and lastmodified
