+++ Slim/Http/Caching/ResourceMapper/Etag.php
@@ -1,0 +1,6 @@
+<?php
+namespace Slim\Http\Caching\ResourceMapper;
+
+class ResourceMapper_Etag extends Base {
+
+}
